requirements.txt
sentence-transformers
python-docx
pandas
numpy
torch
fastapi
uvicorn
sentence-transformers
numpy
python-docx
requests